+++ packages/react-native-bridge/ios/RNReactNativeGutenbergBridge.swift
@@ -130,6 +130,13 @@ public class RNReactNativeGutenbergBridge: RCTEventEmitter {
     }
 
     @objc
+    func setFeaturedImage(_ mediaID: Int32) {
+        DispatchQueue.main.async {
+            self.delegate?.gutenbergDidRequestToSetFeaturedImage(for: mediaID)
+        }
+    }
+
+    @objc
     func editorDidLayout() {
         DispatchQueue.main.async {
             self.delegate?.gutenbergDidLayout()
@@ -373,6 +380,7 @@ extension RNReactNativeGutenbergBridge {
         case setTitle
         case toggleHTMLMode
         case updateHtml
+        case featuredImageIdNativeUpdated
         case mediaUpload
         case setFocusOnTitle
         case mediaAppend